JournalDeleteAction

+++ app/Actions/Sale/Item/DeleteAction.php
@@ -4,6 +4,7 @@ namespace App\Actions\Sale\Item;
 
 use App\Actions\Sale\StockUpdateAction;
 use App\Models\SaleItem;
+use Exception;
 use Illuminate\Support\Facades\Auth;
 
 class DeleteAction
@@ -13,13 +14,16 @@ class DeleteAction
         try {
             $model = SaleItem::find($id);
             if (! $model) {
-                throw new \Exception("Resource not found with the specified ID: $id.", 1);
+                throw new Exception("Resource not found with the specified ID: $id.", 1);
             }
             if ($model->sale->status == 'completed') {
+                if (! Auth::user()->can('sale.edit completed')) {
+                    throw new Exception("You don't have permission to delete it.", 1);
+                }
                 (new StockUpdateAction())->singleItemDelete($model, $model->sale, 'delete_item', Auth::id());
             }
             if (! $model->delete()) {
-                throw new \Exception('Oops! Something went wrong while deleting the SaleItem. Please try again.', 1);
+                throw new Exception('Oops! Something went wrong while deleting the SaleItem. Please try again.', 1);
             }
             $return['success'] = true;
             $return['message'] = 'Successfully Update SaleItem';
